Fix Bugs, get it
