+++ src/widgets/Header/Header.tsx
@@ -4,7 +4,7 @@ import Link from 'next/link';
 import { FC } from 'react';
 import { ShoppingCart, Heart } from 'lucide-react';
 
-import { useAppContext } from '@/context';
+import { useAppContext } from '@/shared/lib/context';
 import { LogoSVG, SearchSVG, UserSVG } from '@/shared/assets/icons';
 
 import { paths } from '@/shared/config/roures';
