+++ src/widgets/ProductCard/ProductCard.tsx
@@ -1,10 +1,10 @@
 'use client';
 
-import { useAppContext } from '@/context';
-import { useProductActions } from '@/hooks';
+import { useAppContext } from '@/shared/lib/context';
 import Image, { StaticImageData } from 'next/image';
 import { Button } from '@/shared/components';
 import { Eye, Heart, ShoppingCart } from 'lucide-react';
+import { useProductActions } from '@/shared/lib/hooks';
 
 import { FC } from 'react';
